Extract ConsoleListViewModel
