+++ Sources/PulseUI/Features/Console/ConsoleViewModel.swift
@@ -7,14 +7,13 @@ import Pulse
 import Combine
 import SwiftUI
 
-final class ConsoleViewModel: NSObject, NSFetchedResultsControllerDelegate, ObservableObject {
+#warning("container for ConsoleVieWModl that doesn't refresh itself")
+
+final class ConsoleViewModel: ObservableObject {
     let title: String
     let isNetworkOnly: Bool
     let store: LoggerStore
-
-    @Published private(set) var visibleEntities: ArraySlice<NSManagedObject> = []
-    @Published private(set) var entities: [NSManagedObject] = []
-    let entitiesSubject = CurrentValueSubject<[NSManagedObject], Never>([])
+    let list: ConsoleListViewModel
 
 #if os(iOS)
     let insightsViewModel: InsightsViewModel
@@ -28,61 +27,41 @@ final class ConsoleViewModel: NSObject, NSFetchedResultsControllerDelegate, Obse
     let searchBarViewModel: ConsoleSearchBarViewModel
     let searchCriteriaViewModel: ConsoleSearchCriteriaViewModel
 
+    #warning("reimplement this")
     var toolbarTitle: String {
         let suffix = mode == .network ? "Requests" : "Messages"
-        return "\(entities.count) \(suffix)"
+        return "\(0) \(suffix)"
     }
 
-    // Sorting/Grouping
-    @Published var messageSortBy: ConsoleMessageSortBy = .dateCreated
-    @Published var taskSortBy: ConsoleTaskSortBy = .dateCreated
-    @Published var order: ConsoleOrdering = .descending
-    @Published var messageGroupBy: ConsoleMessageGroupBy = .plain
-    @Published var taskGroupBy: ConsoleTaskGroupBy = .plain
-
     // Filters
-    @Published var mode: Mode
+    @Published var mode: ConsoleMode
     @Published var isOnlyErrors = false
     @Published var filterTerm: String = ""
     @Published var isShowingFilters = false
 
     var onDismiss: (() -> Void)?
 
-    /// This exist strickly to workaround List performance issues
-    private var scrollPosition: ScrollPosition = .nearTop
-    private var visibleEntityCountLimit = fetchBatchSize
-    private var visibleObjectIDs: Set<NSManagedObjectID> = []
-
-    private var controller: NSFetchedResultsController<NSManagedObject>?
     private var isViewVisible = false
     private var cancellables: [AnyCancellable] = []
 
-    enum Mode {
-        case messages, network
-    }
-
-    init(store: LoggerStore, mode: Mode = .messages) {
+    init(store: LoggerStore, mode: ConsoleMode = .messages) {
         self.title = mode == .network ? "Network" : "Console"
         self.store = store
         self.mode = mode
         self.isNetworkOnly = mode == .network
-        self.order = .descending
 
+        self.searchCriteriaViewModel = ConsoleSearchCriteriaViewModel(store: store)
         self.searchBarViewModel = ConsoleSearchBarViewModel()
-        self.searchCriteriaViewModel = ConsoleSearchCriteriaViewModel(store: store, entities: entitiesSubject)
+        self.list = ConsoleListViewModel(store: store, mode: mode, criteria: searchCriteriaViewModel)
 
 #if os(iOS)
         self.insightsViewModel = InsightsViewModel(store: store)
         if #available(iOS 15, *) {
-            self._searchViewModel = ConsoleSearchViewModel(entities: entitiesSubject, store: store, searchBar: searchBarViewModel)
+            self._searchViewModel = ConsoleSearchViewModel(entities: list.entitiesSubject, store: store, searchBar: searchBarViewModel)
         }
 #endif
 
-        super.init()
-
-        $entities.sink { [entitiesSubject] in
-            entitiesSubject.send($0)
-        }.store(in: &cancellables)
+        searchCriteriaViewModel.bind(list.$entities)
 
         $filterTerm
             .dropFirst()
@@ -94,18 +73,14 @@ final class ConsoleViewModel: NSObject, NSFetchedResultsControllerDelegate, Obse
         searchCriteriaViewModel.$criteria
             .dropFirst()
             .throttle(for: 0.5, scheduler: DispatchQueue.main, latest: true)
-            .sink { [weak self] _ in self?.refreshNow() }
+            .sink { [weak self] _ in self?.refreshList() }
             .store(in: &cancellables)
 
         $isOnlyErrors.receive(on: DispatchQueue.main).dropFirst().sink { [weak self] _ in
-            self?.refreshNow()
+            self?.refreshList()
         }.store(in: &cancellables)
 
-        $order.dropFirst().receive(on: DispatchQueue.main).sink { [weak self] _ in
-            self?.prepareFetchController()
-        }.store(in: &cancellables)
-
-        prepareFetchController()
+        list.refreshController()
     }
 
     // MARK: Mode
@@ -115,35 +90,16 @@ final class ConsoleViewModel: NSObject, NSFetchedResultsControllerDelegate, Obse
         case .messages: mode = .network
         case .network: mode = .messages
         }
-        prepareFetchController()
-    }
-
-    private func prepareFetchController() {
-        searchCriteriaViewModel.mode = mode
-
-        let request = makeFetchRequest(for: mode, order: order)
-        controller = NSFetchedResultsController(fetchRequest: request, managedObjectContext: store.viewContext, sectionNameKeyPath: nil, cacheName: nil)
-        controller?.delegate = self
-
-        refreshNow()
-    }
-
-    // MARK: Appearance
-
-    func onAppear() {
-        isViewVisible = true
-        reloadMessages(isMandatory: true)
-    }
-
-    func onDisappear() {
-        isViewVisible = false
+        list.refreshController()
     }
 
     // MARK: Refresh
 
-    private func refreshNow() {
+    private func refreshList() {
         // important: order
         refresh(filterTerm: filterTerm)
+
+#warning("reomplement")
 #if os(iOS)
         if #available(iOS 15, *) {
             searchViewModel.refreshNow()
@@ -151,142 +107,18 @@ final class ConsoleViewModel: NSObject, NSFetchedResultsControllerDelegate, Obse
 #endif
     }
 
+    #warning("remove")
     private func refresh(filterTerm: String) {
-        // Search messages
-        guard let controller = controller else {
-            return assertionFailure()
-        }
-        switch mode {
-        case .messages:
-            controller.fetchRequest.predicate = ConsoleSearchCriteria.makeMessagePredicates(criteria: searchCriteriaViewModel.criteria, isOnlyErrors: isOnlyErrors, filterTerm: filterTerm)
-        case .network:
-            controller.fetchRequest.predicate = ConsoleSearchCriteria.makeNetworkPredicates(criteria: searchCriteriaViewModel.criteria, isOnlyErrors: isOnlyErrors, filterTerm: filterTerm)
-        }
-        try? controller.performFetch()
-
-        reloadMessages()
+        list.refresh()
     }
-
-    // MARK: - NSFetchedResultsControllerDelegate
-
-    func controller(_ controller: NSFetchedResultsController<NSFetchRequestResult>, didChangeContentWith diff: CollectionDifference<NSManagedObjectID>) {
-        if isViewVisible {
-            withAnimation {
-                reloadMessages(isMandatory: false)
-            }
-        } else {
-            entities = self.controller?.fetchedObjects ?? []
-        }
-    }
-
-    private func reloadMessages(isMandatory: Bool = true) {
-        entities = controller?.fetchedObjects ?? []
-        if isMandatory || scrollPosition == .nearTop {
-            refreshVisibleEntities()
-        }
-    }
-
-    private enum ScrollPosition {
-        case nearTop
-        case middle
-        case nearBottom
-    }
-
-    private func refreshVisibleEntities() {
-        visibleEntities = entities.prefix(visibleEntityCountLimit)
-    }
-
-    func onDisappearCell(with objectID: NSManagedObjectID) {
-        visibleObjectIDs.remove(objectID)
-        refreshScrollPosition()
-    }
-
-    func onAppearCell(with objectID: NSManagedObjectID) {
-        visibleObjectIDs.insert(objectID)
-        refreshScrollPosition()
-    }
-
-    private func refreshScrollPosition() {
-        let scrollPosition: ScrollPosition
-        if visibleEntities.prefix(5).map(\.objectID).contains(where: visibleObjectIDs.contains) {
-            scrollPosition = .nearTop
-        } else if visibleEntities.suffix(5).map(\.objectID).contains(where: visibleObjectIDs.contains) {
-            scrollPosition = .nearBottom
-        } else {
-            scrollPosition = .middle
-        }
-
-        if scrollPosition != self.scrollPosition {
-            self.scrollPosition = scrollPosition
-            switch scrollPosition {
-            case .nearTop:
-                visibleEntityCountLimit = fetchBatchSize // Reset
-                refreshVisibleEntities()
-            case .middle:
-                break // Don't reload: too expensive and ruins gestures
-            case .nearBottom:
-                visibleEntityCountLimit += fetchBatchSize
-                refreshVisibleEntities()
-            }
-        }
-    }
-
 
     // MARK: - Sharing
 
     func prepareForSharing(as output: ShareOutput, _ completion: @escaping (ShareItems?) -> Void) {
-        ShareService.share(entities, store: store, as: output, completion)
+        ShareService.share(list.entities, store: store, as: output, completion)
     }
 }
 
-private func makeFetchRequest(for mode: ConsoleViewModel.Mode, order: ConsoleOrdering) -> NSFetchRequest<NSManagedObject> {
-    let request: NSFetchRequest<NSManagedObject>
-    let isAscending = order == .descending
-    switch mode {
-    case .messages:
-        request = .init(entityName: "\(LoggerMessageEntity.self)")
-        request.relationshipKeyPathsForPrefetching = ["request"]
-        request.sortDescriptors = [NSSortDescriptor(keyPath: \LoggerMessageEntity.createdAt, ascending: isAscending)]
-    case .network:
-        request = .init(entityName: "\(NetworkTaskEntity.self)")
-        request.sortDescriptors = [NSSortDescriptor(keyPath: \NetworkTaskEntity.createdAt, ascending: isAscending)]
-    }
-    request.fetchBatchSize = fetchBatchSize
-    return request
-}
-
-private let fetchBatchSize = 100
-
-enum ConsoleOrdering: String, CaseIterable {
-    case descending = "Descending"
-    case ascending = "Ascending"
-}
-
-enum ConsoleMessageSortBy: String, CaseIterable {
-    case dateCreated = "Date Created"
-}
-
-enum ConsoleTaskSortBy: String, CaseIterable {
-    case dateCreated = "Date Created"
-    case duration = "Duration"
-    case requestSize = "Request Size"
-    case responseSize = "Response Size"
-}
-
-enum ConsoleMessageGroupBy: String, CaseIterable {
-    case plain = "None"
-    case label = "Label"
-    case level = "Level"
-    case file = "File"
-}
-
-enum ConsoleTaskGroupBy: String, CaseIterable {
-    case plain = "None"
-    case url = "URL"
-    case host = "Host"
-    case method = "Method"
-    case scheme = "Scheme"
-    case taskType = "Task Type"
-    case statusCode = "Status Code"
-    case errorCode = "Error Code"
+enum ConsoleMode {
+    case messages, network
 }
